+++ packages/plugin-text/src/factory/index.tsx
@@ -46,13 +46,15 @@ export const createTextPlugin = (
       }
       return false
     },
-    isEmpty: (state: StateDescriptorValueType<typeof textState>) => {
-      const value = Value.fromJSON(state)
-      return (
-        value.document.text === '' &&
-        value.document.nodes.size === 1 &&
-        value.document.getTexts().size === 1
-      )
-    }
+    isEmpty: (state: StateDescriptorValueType<typeof textState>) =>
+      isValueEmpty(Value.fromJSON(state))
   }
+}
+
+export function isValueEmpty(value: Value) {
+  return (
+    value.document.text === '' &&
+    value.document.nodes.size === 1 &&
+    value.document.getTexts().size === 1
+  )
 }