+++ scripts/changelog.ts
@@ -171,6 +171,9 @@ async function exec(): Promise<void> {
       ]
     },
     {
+      tagName: 'v0.5.0',
+      name: '0.5.0',
+      date: '2019-05-07',
       breakingChanges: [
         '**core**. Mark `name` as required in `StatelessEditorProps`',
         '**plugin-image**. Image-plugin has new required `maxWidth` - state property',
@@ -205,6 +208,13 @@ async function exec(): Promise<void> {
       internal: [
         '**demo**. Log errors only',
         '**demo**. Add additional examples for plugin-usage in Storybook'
+      ]
+    },
+    {
+      added: [
+        '**plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.',
+        '**core**. Pass plugin props and name of parents to childs too',
+        '**plugin-text**. Leave nested Text-Plugins on second Enter key press'
       ]
     }
   ])